debug: move arch_get_stack_*() functions to a separate header
arch_get_stack_*() functions are only used in panic.c. Move them to
a separate header to reduce dependencies.
