+++ src/include/sof/debug/backtrace.h
@@ -1,0 +1,11 @@
+/* SPDX-License-Identifier: BSD-3-Clause
+ *
+ * Copyright(c) 2020 Intel Corporation. All rights reserved.
+ */
+
+#ifndef __SOF_DEBUG_BACKTRACE_H__
+#define __SOF_DEBUG_BACKTRACE_H__
+
+#include <arch/debug/backtrace.h>
+
+#endif
